Add verbosity level to log messages
This commit adds a verbosity level to log messages. The verbosity level
can be configure in the .cnf file or else Note will be assumed by
default. This is to avoid the scheduler from flooding the logs with
messages that indicate the scheduler is running as expected.

There are 4 log levels:
- Info
- Note
- Warning
- Error

The last two are obvious, but the first two are as follows:
- Info: This is just and information, the action that follow this will
not change the state of the proxy.
- Note: This is a note to the user that the action that follows this
will most likely change the state of the proxy.

Closes #16