Force cache creation to set default schema
Force connection to set the default schema before attempting to
create a cache. We use one single connection per host. During cache
creation dry-run we enforce the default schema to be set, but During
actual cache creation we do not. In case a client has more than one
Readyset server, we will not execute the dry-run on the second server
onwards. This will cause the cache creation to fail as the default
schema is not set.

This commit forces the default schema to be set before attempting to
create a cache.

Fixes: #31