Force cache creation to set default schema
Force connection to set the default schema before attempting to
create a cache. We use one single connection per host. During cache
creation dry-run we enforce the default schema to be set, but During
actual cache creation we do not. In case a client has more than one
schema, we might end up attempting to create the cache in the wrong
schema or cache creation might fail due to the table not being found.