Consolidate config defaults
Use #[serde(default)] to provide default values for missing
configuration options during deserialization.  This typically makes use
of the type's Default trait, but we also specify a couple default
functions for standard types when a custom default is desired.

Also make number_of_queries optional to bring its semantics into
alignment with the documentation in the README.md.  I think it was
accidentally specified as a required option.