Allow configuration for PostgreSQL
This patch adds a new and optional configuration option, database_type,
to the set of options available in the scheduler's config.  Valid values
are "mysql" and "postgresql", and if the option is omitted, "mysql" as
treated as the default.

Note that this patch does not yet introduce PostgreSQL support, but
provides the initial restructuring of the scheduler.

Abstract usage of an SQL connection into a new module, sql_connection,
and a new enum representing a connection, SQLConnection, which
represents either a MySQL or a PostgreSQL connection.  Also replace all
uses of mysql::Conn with the new SQLConnection enum, which acts as a
thin wrapper.