Add initial PostgreSQL support
This patch adds PostgreSQL functionality to our ProxySQL scheduler,
filling out the second half of our sql_connection module.  For now, we
support health checks in PostgreSQL, but not query discovery.

Be sure to set database_type = "postgresql" in the scheduler's config to
use the PostgreSQL flavor of the scheduler.